Formatting Output Using Library Functions.
Added a C++ source file to illustrate the use of ios flags in combination with the library functions like setprecision() and setw() to manipulate the format of output.
The following ios flags have been used:
(1) left;
(2) right;
(3) fixed;
(4) showpos;
(5) showpoint;
(6) uppercase; and
(7) scientific.

unsetf() method of 'cout' class and resetiosflags() function is also demonstrated to reset the formatting.